organize icon files, minor refactoring
