+++ src/components/navigation/Navigation.stories.mdx
@@ -1,6 +1,6 @@
 import { Meta, Story, Preview } from "@storybook/addon-docs/blocks";
 import { Navigation } from ".";
-import Decos from "../../utils/decos.tsx";
+import { Decos } from "./decos.tsx";
 import { SVGIcon } from "../svgicon";
 import { Box } from "../box";
 import { Text } from "../text";
