+++ src/components/svgicon/icons/aws-resource.tsx
