+++ src/components/svgicon/icons/cloud-provider.tsx
