+++ src/components/svgicon/icons/index.ts
@@ -1,0 +1,8 @@
+export { awsIcons } from "./aws-resource";
+export { cloudIcons } from "./cloud-provider";
+export { Decos } from "../../navigation/decos";
+export { klarityIcons } from "./klarity-resource";
+export { menuIcons } from "./menu";
+export { osComponentsIcons } from "./os-components";
+export { osProviderIcons } from "./os-provider";
+export { uiIcons } from "./ui";
