+++ src/components/svgicon/icons/klarity-resource.tsx
