--- src/utils/icons.ts
@@ -1,251 +1,0 @@
-export const icons = {
-  STATUS_ENABLED: `<?xml version="1.0" encoding="utf-8"?>
-    <!-- Generator: Adobe Illustrator 24.0.3, SVG Export Plug-In . SVG Version: 6.00 Build 0)  -->
-    <svg version="1.1" id="Warstwa_1" xmlns="http://www.w3.org/2000/svg" xmlns:xlink="http://www.w3.org/1999/xlink" x="0px" y="0px"
-         viewBox="0 0 24 24" style="enable-background:new 0 0 24 24;" xml:space="preserve">
-    <style type="text/css">
-        .st0{fill:#27AE60;}
-        .st1{fill:#45455B;}
-    </style>
-    <g transform="matrix(1 0 0 1 -7 -11 )">
-        <path class="st0" d="M23.93,28.35c0-2.72-2.21-4.93-4.93-4.93s-4.93,2.21-4.93,4.93s2.21,4.93,4.93,4.93S23.93,31.07,23.93,28.35z
-             M21.77,27.17l-3.16,3.33l-2.39-2.33l0.98-1.01l1.37,1.33l2.18-2.3L21.77,27.17z"/>
-    </g>
-    <g>
-        <path class="st1" d="M24,12.44c0-3.3-2.68-5.98-5.98-5.98c-0.31-1.28-1-2.45-2.01-3.33c-1.18-1.04-2.69-1.61-4.26-1.61
-            c-3.55,0-6.44,2.89-6.44,6.44c0,0.06,0,0.11,0,0.17c-0.05,0-0.11,0-0.16,0C2.31,8.12,0,10.43,0,13.27s2.31,5.15,5.15,5.15h0.61
-            c-0.06-0.35-0.09-0.71-0.09-1.08c0-1.69,0.66-3.28,1.85-4.47s2.78-1.85,4.47-1.85s3.28,0.66,4.47,1.85s1.85,2.78,1.85,4.47
-            c0,0.36-0.03,0.72-0.09,1.07C21.43,18.3,24,15.66,24,12.44z"/>
-    </g>
-    </svg>`,
-  STATUS_DISABLED: `<?xml version="1.0" encoding="utf-8"?>
-    <!-- Generator: Adobe Illustrator 24.0.3, SVG Export Plug-In . SVG Version: 6.00 Build 0)  -->
-    <svg version="1.1" id="Warstwa_1" xmlns="http://www.w3.org/2000/svg" xmlns:xlink="http://www.w3.org/1999/xlink" x="0px" y="0px"
-         viewBox="0 0 24 24" style="enable-background:new 0 0 24 24;" xml:space="preserve">
-    <style type="text/css">
-        .st0{fill:#45455B;}
-        .st1{fill:#E74C3C;}
-    </style>
-    <g>
-        <path class="st0" d="M24,12.44c0-3.3-2.68-5.98-5.98-5.98c-0.31-1.28-1-2.45-2.01-3.33c-1.18-1.04-2.69-1.61-4.26-1.61
-            c-3.55,0-6.44,2.89-6.44,6.44c0,0.06,0,0.11,0,0.17c-0.05,0-0.11,0-0.16,0C2.31,8.12,0,10.43,0,13.27s2.31,5.15,5.15,5.15h0.61
-            c-0.06-0.35-0.09-0.71-0.09-1.08c0-1.69,0.66-3.28,1.85-4.47s2.78-1.85,4.47-1.85s3.28,0.66,4.47,1.85s1.85,2.78,1.85,4.47
-            c0,0.36-0.03,0.72-0.09,1.07C21.43,18.3,24,15.66,24,12.44z"/>
-    </g>
-    <g transform="matrix(1 0 0 1 -7 -11 )">
-        <path class="st1" d="M23.93,28.35c0-2.72-2.21-4.93-4.93-4.93s-4.93,2.21-4.93,4.93s2.21,4.93,4.93,4.93S23.93,31.07,23.93,28.35z
-             M19.98,28.35l1.26,1.26l-0.99,0.99l-1.26-1.26l-1.26,1.26l-0.99-0.99L18,28.35l-1.26-1.26l0.99-0.99l1.26,1.26l1.26-1.26
-            l0.99,0.99L19.98,28.35z"/>
-    </g>
-    </svg>`,
-  STATUS_WARNING: `<?xml version="1.0" encoding="utf-8"?>
-  <!-- Generator: Adobe Illustrator 24.0.3, SVG Export Plug-In . SVG Version: 6.00 Build 0)  -->
-  <svg version="1.1" id="Capa_1" xmlns="http://www.w3.org/2000/svg" xmlns:xlink="http://www.w3.org/1999/xlink" x="0px" y="0px"
-     viewBox="0 0 24 24" style="enable-background:new 0 0 24 24;" xml:space="preserve">
-  <style type="text/css">
-    .st0{fill:#45455B;}
-    .st1{fill:#F39C12;}
-    .st2{fill:#FFFFFF;}
-  </style>
-  <g>
-    <g>
-      <path class="st0" d="M18,6.6c-0.3-1.3-1-2.4-2-3.3c-1.2-1-2.7-1.6-4.3-1.6c-3.6,0-6.4,2.9-6.4,6.4c0,0.1,0,0.1,0,0.2
-        c-0.1,0-0.1,0-0.2,0C2.3,8.2,0,10.5,0,13.4s2.3,5.2,5.2,5.2h0.6c-0.1-0.4-0.1-0.7-0.1-1.1c0-1.7,0.7-3.3,1.9-4.5s2.8-1.9,4.5-1.9
-        c1.7,0,3.3,0.7,4.5,1.9s1.9,2.8,1.9,4.5c0,0.4,0,0.7-0.1,1.1c3.2-0.1,5.8-2.8,5.8-6C24,9.2,21.3,6.6,18,6.6z"/>
-    </g>
-  </g>
-  <circle class="st1" cx="12" cy="17.5" r="4.9"/>
-  <g>
-    <path class="st2" d="M12.9,14.5c0,1.3-0.1,2.2-0.1,2.8c-0.1,0.6-0.1,1-0.1,1.2h-1.3c0-0.2-0.1-0.7-0.1-1.2s-0.1-1.5-0.1-2.8H12.9z
-       M12.9,20.8h-1.7v-1.5h1.7V20.8z"/>
-  </g>
-  </svg>`,
-  AWS: `<svg xmlns="http://www.w3.org/2000/svg" viewBox="0 0 48 48" width="48px" height="48px"><path fill="#252f3e" d="M13.527,21.529c0,0.597,0.064,1.08,0.176,1.435c0.128,0.355,0.287,0.742,0.511,1.161 c0.08,0.129,0.112,0.258,0.112,0.371c0,0.161-0.096,0.322-0.303,0.484l-1.006,0.677c-0.144,0.097-0.287,0.145-0.415,0.145 c-0.16,0-0.319-0.081-0.479-0.226c-0.224-0.242-0.415-0.5-0.575-0.758c-0.16-0.274-0.319-0.58-0.495-0.951 c-1.245,1.483-2.81,2.225-4.694,2.225c-1.341,0-2.411-0.387-3.193-1.161s-1.181-1.806-1.181-3.096c0-1.37,0.479-2.483,1.453-3.321 s2.267-1.258,3.911-1.258c0.543,0,1.102,0.048,1.692,0.129s1.197,0.21,1.836,0.355v-1.177c0-1.225-0.255-2.08-0.75-2.58 c-0.511-0.5-1.373-0.742-2.602-0.742c-0.559,0-1.133,0.064-1.724,0.21c-0.591,0.145-1.165,0.322-1.724,0.548 c-0.255,0.113-0.447,0.177-0.559,0.21c-0.112,0.032-0.192,0.048-0.255,0.048c-0.224,0-0.335-0.161-0.335-0.5v-0.79 c0-0.258,0.032-0.451,0.112-0.564c0.08-0.113,0.224-0.226,0.447-0.339c0.559-0.29,1.229-0.532,2.012-0.726 c0.782-0.21,1.612-0.306,2.49-0.306c1.9,0,3.289,0.435,4.183,1.306c0.878,0.871,1.325,2.193,1.325,3.966v5.224H13.527z M7.045,23.979c0.527,0,1.07-0.097,1.644-0.29c0.575-0.193,1.086-0.548,1.517-1.032c0.255-0.306,0.447-0.645,0.543-1.032 c0.096-0.387,0.16-0.855,0.16-1.403v-0.677c-0.463-0.113-0.958-0.21-1.469-0.274c-0.511-0.064-1.006-0.097-1.501-0.097 c-1.07,0-1.852,0.21-2.379,0.645s-0.782,1.048-0.782,1.854c0,0.758,0.192,1.322,0.591,1.709 C5.752,23.786,6.311,23.979,7.045,23.979z M19.865,25.721c-0.287,0-0.479-0.048-0.607-0.161c-0.128-0.097-0.239-0.322-0.335-0.629 l-3.752-12.463c-0.096-0.322-0.144-0.532-0.144-0.645c0-0.258,0.128-0.403,0.383-0.403h1.565c0.303,0,0.511,0.048,0.623,0.161 c0.128,0.097,0.223,0.322,0.319,0.629l2.682,10.674l2.49-10.674c0.08-0.322,0.176-0.532,0.303-0.629 c0.128-0.097,0.351-0.161,0.639-0.161h1.277c0.303,0,0.511,0.048,0.639,0.161c0.128,0.097,0.239,0.322,0.303,0.629l2.522,10.803 l2.762-10.803c0.096-0.322,0.208-0.532,0.319-0.629c0.128-0.097,0.335-0.161,0.623-0.161h1.485c0.255,0,0.399,0.129,0.399,0.403 c0,0.081-0.016,0.161-0.032,0.258s-0.048,0.226-0.112,0.403l-3.847,12.463c-0.096,0.322-0.208,0.532-0.335,0.629 s-0.335,0.161-0.607,0.161h-1.373c-0.303,0-0.511-0.048-0.639-0.161c-0.128-0.113-0.239-0.322-0.303-0.645l-2.474-10.4 L22.18,24.915c-0.08,0.322-0.176,0.532-0.303,0.645c-0.128,0.113-0.351,0.161-0.639,0.161H19.865z M40.379,26.156 c-0.83,0-1.66-0.097-2.458-0.29c-0.798-0.193-1.421-0.403-1.836-0.645c-0.255-0.145-0.431-0.306-0.495-0.451 c-0.064-0.145-0.096-0.306-0.096-0.451v-0.822c0-0.339,0.128-0.5,0.367-0.5c0.096,0,0.192,0.016,0.287,0.048 c0.096,0.032,0.239,0.097,0.399,0.161c0.543,0.242,1.133,0.435,1.756,0.564c0.639,0.129,1.261,0.193,1.9,0.193 c1.006,0,1.788-0.177,2.331-0.532c0.543-0.355,0.83-0.871,0.83-1.532c0-0.451-0.144-0.822-0.431-1.129 c-0.287-0.306-0.83-0.58-1.612-0.838l-2.315-0.726c-1.165-0.371-2.027-0.919-2.554-1.645c-0.527-0.709-0.798-1.499-0.798-2.338 c0-0.677,0.144-1.274,0.431-1.79s0.671-0.967,1.149-1.322c0.479-0.371,1.022-0.645,1.66-0.838C39.533,11.081,40.203,11,40.906,11 c0.351,0,0.718,0.016,1.07,0.064c0.367,0.048,0.702,0.113,1.038,0.177c0.319,0.081,0.623,0.161,0.91,0.258s0.511,0.193,0.671,0.29 c0.224,0.129,0.383,0.258,0.479,0.403c0.096,0.129,0.144,0.306,0.144,0.532v0.758c0,0.339-0.128,0.516-0.367,0.516 c-0.128,0-0.335-0.064-0.607-0.193c-0.91-0.419-1.932-0.629-3.065-0.629c-0.91,0-1.628,0.145-2.123,0.451 c-0.495,0.306-0.75,0.774-0.75,1.435c0,0.451,0.16,0.838,0.479,1.145c0.319,0.306,0.91,0.613,1.756,0.887l2.267,0.726 c1.149,0.371,1.98,0.887,2.474,1.548s0.734,1.419,0.734,2.257c0,0.693-0.144,1.322-0.415,1.87 c-0.287,0.548-0.671,1.032-1.165,1.419c-0.495,0.403-1.086,0.693-1.772,0.903C41.943,26.043,41.193,26.156,40.379,26.156z"/><path fill="#f90" d="M43.396,33.992c-5.252,3.918-12.883,5.998-19.445,5.998c-9.195,0-17.481-3.434-23.739-9.142 c-0.495-0.451-0.048-1.064,0.543-0.709c6.769,3.966,15.118,6.369,23.755,6.369c5.827,0,12.229-1.225,18.119-3.741 C43.508,32.364,44.258,33.347,43.396,33.992z M45.583,31.477c-0.671-0.871-4.438-0.419-6.146-0.21 c-0.511,0.064-0.591-0.387-0.128-0.726c3.001-2.128,7.934-1.516,8.509-0.806c0.575,0.726-0.16,5.708-2.969,8.094 c-0.431,0.371-0.846,0.177-0.655-0.306C44.833,35.927,46.254,32.331,45.583,31.477z"/></svg>`,
-  AZURE: `<svg xmlns="http://www.w3.org/2000/svg" viewBox="0 0 48 48" width="48px" height="48px"><path fill="#035bda" d="M46 40L29.317 10.852 22.808 23.96 34.267 37.24 13 39.655zM13.092 18.182L2 36.896 11.442 35.947 28.033 5.678z"/></svg>`,
-  GCP: `<svg xmlns="http://www.w3.org/2000/svg" viewBox="0 0 48 48" width="48px" height="48px"><path fill="#CFD8DC" d="M30,13H18l-7,11l7,11h12l7-11L30,13z M24,29c-2.8,0-5-2.2-5-5s2.2-5,5-5s5,2.2,5,5S26.8,29,24,29z"/><path fill="#2196F3" d="M34.8,7.2C34.3,6.5,33.5,6,32.6,6H15.4c-0.9,0-1.7,0.5-2.2,1.2l-5.8,9.5L12,24l6.3-10h20.6L34.8,7.2z M16,11c-0.6,0-1-0.4-1-1s0.4-1,1-1c0.6,0,1,0.4,1,1S16.6,11,16,11z M32,11c-0.6,0-1-0.4-1-1s0.4-1,1-1c0.6,0,1,0.4,1,1S32.6,11,32,11z"/><path fill="#FFC107" d="M18.3,34L7.4,16.8l-3.6,5.9c-0.5,0.8-0.5,1.9,0,2.7l9.4,15.4c0.5,0.8,1.3,1.2,2.2,1.2h9.1l5.1-8H18.3z M8,25c-0.6,0-1-0.4-1-1s0.4-1,1-1c0.6,0,1,0.4,1,1S8.6,25,8,25z M16,39c-0.6,0-1-0.4-1-1s0.4-1,1-1c0.6,0,1,0.4,1,1S16.6,39,16,39z"/><path fill="#1976D2" d="M7.4 16.8L12 24 14.4 20.3 8.5 15z"/><path fill="#F9A825" d="M24.6 42L29.7 34 24.9 34 22.6 42z"/><path fill="#DD2C00" d="M44.2,22.7L38.9,14h-9.2L36,24L24.6,42h8c0.9,0,1.7-0.5,2.2-1.2l9.5-15.4C44.8,24.5,44.8,23.5,44.2,22.7z M32,39c-0.6,0-1-0.4-1-1s0.4-1,1-1c0.6,0,1,0.4,1,1S32.6,39,32,39z M40,25c-0.6,0-1-0.4-1-1s0.4-1,1-1c0.6,0,1,0.4,1,1S40.6,25,40,25z"/><path fill="#BF360C" d="M38.9 14L29.7 14 32.2 17.9 40.1 16z"/></svg>`,
-  WINDOWS: `<?xml version="1.0" encoding="iso-8859-1"?>
-<!-- Generator: Adobe Illustrator 19.0.0, SVG Export Plug-In . SVG Version: 6.00 Build 0)  -->
-<svg version="1.1" id="Capa_1" xmlns="http://www.w3.org/2000/svg" xmlns:xlink="http://www.w3.org/1999/xlink" x="0px" y="0px"
-	 viewBox="0 0 512 512" style="enable-background:new 0 0 512 512;" xml:space="preserve">
-<path style="fill:#F44336;" d="M201.302,49.754l-192,24.939C3.949,75.384-0.044,79.963,0,85.36v149.333
-	c0,5.891,4.776,10.667,10.667,10.667h192c5.891,0,10.667-4.776,10.667-10.667V60.314c-0.004-3.064-1.325-5.978-3.627-8
-	C207.388,50.32,204.339,49.391,201.302,49.754z"/>
-<path style="fill:#4CAF50;" d="M508.374,13.36c-2.286-2.06-5.35-3.032-8.405-2.667l-256,33.387
-	c-5.352,0.691-9.346,5.27-9.301,10.667v179.947c0,5.891,4.776,10.667,10.667,10.667h256c5.891,0,10.667-4.776,10.667-10.667V21.36
-	C511.997,18.296,510.675,15.382,508.374,13.36z"/>
-<path style="fill:#2196F3;" d="M202.667,266.693h-192C4.776,266.693,0,271.469,0,277.36v149.333
-	c-0.044,5.397,3.949,9.976,9.301,10.667l192,25.045c0.455,0.031,0.911,0.031,1.365,0c5.891,0,10.667-4.776,10.667-10.667V277.36
-	C213.334,271.469,208.558,266.693,202.667,266.693z"/>
-<path style="fill:#FFC107;" d="M501.334,266.693h-256c-5.891,0-10.667,4.776-10.667,10.667v179.947
-	c-0.044,5.397,3.949,9.976,9.301,10.667l256,33.387c0.455,0.031,0.911,0.031,1.365,0c5.891,0,10.667-4.776,10.667-10.667v0V277.36
-	C512,271.469,507.225,266.693,501.334,266.693z"/>
-</svg>
-`,
-  LINUX: `<?xml version="1.0" encoding="UTF-8" standalone="no"?> <svg xmlns="http://www.w3.org/2000/svg"
-aria-label="Linux" role="img"
-viewBox="0 0 512 512" fill="#333"><rect
-width="512" height="512"
-rx="15%"
-fill="#fff"/><g transform="matrix(2 0 0 2 256 256)"><path d="M-32-25c-3 7-24 29-22 51 8 92 36 30 78 53 0 0 75-42 15-110-17-24-2-43-13-59s-30-17-44-2 6 37-14 67"/><path d="M42 21s9-18-8-31c16 17 6 32 6 32h-3C36-13 27 6 14-56 29-73 0-88 0-60h-9c1-24-20-12-8 5-1 37-23 52-23 78-7-18 6-32 6-32s-18 15-7 37 31 17 17 27c22 15 56 5 55-27 1-8 22-5 24-3s-3-4-13-4m-56-78c-7-2-5-11-2-11s8 7 2 11m19 1c-5-7-1-14 4-13s5 13-4 13" fill="#eee"/><g fill="#fc2" stroke="#333" stroke-width="1"><path d="M-41 31l21 30c11 7 5 35-25 21-17-5-31-4-33-13s4-10 3-14c-4-22 14-11 19-22s5-16 15-2M71 45c-4-6 0-17-14-16-6 12-23 24-24 0-10 0-3 24-7 35-9 27 17 29 28 16l26-18c2-3 5-6-9-17m-92-92c-3-6 11-14 16-14s12 4 19 6 4 9 2 10S3-35-5-35s-10-8-16-12"/><path d="M-21-48c8 6 17 11 35-3"/></g><path d="M-10-54c-2 0 1-2 2-1m7 1c1-1-1-2-3-1"/></g></svg>`,
-  AWS_ASG: `<svg xmlns="http://www.w3.org/2000/svg" xmlns:xlink="http://www.w3.org/1999/xlink" viewBox="0 0 75 75"><defs><style>.cls-1{fill:url(#OrangeGradient);}.cls-2{fill:#fff;}</style><linearGradient id="OrangeGradient" x1="86.87" y1="-288.75" x2="236.87" y2="-288.75" gradientTransform="translate(127.22 356.13) rotate(-45)" gradientUnits="userSpaceOnUse"><stop offset="0" stop-color="#c8511b"/><stop offset="1" stop-color="#f90"/></linearGradient></defs><title>Amazon-EC2-Auto-Scaling</title><g id="Reference"><rect id="Orange_Gradient" data-name="Orange Gradient" class="cls-1" width="75" height="75"/><g id="Icon_Test" data-name="Icon Test"><path class="cls-2" d="M36.5,15.91v9.58h2V15.91l4.66,4.66,1.41-1.42-6.36-6.36a1,1,0,0,0-1.42,0l-6.36,6.36,1.41,1.42Z"/><path class="cls-2" d="M38.5,59.08V49.49h-2v9.59l-4.66-4.66-1.41,1.42,6.36,6.36a1,1,0,0,0,1.42,0l6.36-6.36-1.41-1.42Z"/><path class="cls-2" d="M44.5,29.65V27.49h-2v2h-2v-2h-2v2h-2v-2h-2v2h-2v-2h-2v2.16a1.7,1.7,0,0,0-.85.84H27.5v2h2v2h-2v2h2v2h-2v2h2v2h-2v2h2.15a1.68,1.68,0,0,0,.85.85v2.15h2v-2h2v2h2v-2h2v2h2v-2h2v2h2V45.34a1.68,1.68,0,0,0,.85-.85H47.5v-2h-2v-2h2v-2h-2v-2h2v-2h-2v-2h2v-2H45.35A1.7,1.7,0,0,0,44.5,29.65Zm-1,13.84h-12v-12h12Z"/><path class="cls-2" d="M18.77,30.49l-1.41,1.42L22,36.49H12.5v2h9.59l-4.73,4.73,1.41,1.41,6.37-6.36a1,1,0,0,0,.29-.71,1,1,0,0,0-.29-.7Z"/><path class="cls-2" d="M62.5,36.49H53.05l4.59-4.58-1.41-1.42-6.37,6.37a1,1,0,0,0-.29.7,1,1,0,0,0,.29.71l6.37,6.36,1.41-1.41-4.73-4.73H62.5Z"/></g></g></svg>`,
-  AWS_EC2: `<svg xmlns="http://www.w3.org/2000/svg" xmlns:xlink="http://www.w3.org/1999/xlink" viewBox="0 0 75 75"><defs><style>.cls-1{fill:url(#OrangeGradient);}.cls-2{fill:#fff;}</style><linearGradient id="OrangeGradient" x1="1131.1" y1="-516.48" x2="1131.1" y2="-366.48" gradientTransform="translate(1149.48 525.13) rotate(-135)" gradientUnits="userSpaceOnUse"><stop offset="0" stop-color="#c8511b"/><stop offset="1" stop-color="#f90"/></linearGradient></defs><title>Amazon-EC2</title><g id="Working"><rect id="Orange_Gradient" data-name="Orange Gradient" class="cls-1" width="75" height="75"/><g id="Icon_Test" data-name="Icon Test"><path class="cls-2" d="M60.49,40.5h-5v-2h5v-24h-24v5h-2v-5a2,2,0,0,1,2-2h24a2,2,0,0,1,2,2v24A2,2,0,0,1,60.49,40.5Z"/><path class="cls-2" d="M38.49,62.5h-24a2,2,0,0,1-2-2v-24a2,2,0,0,1,2-2h5v2h-5v24h24v-5h2v5A2,2,0,0,1,38.49,62.5Z"/><path class="cls-2" d="M53.5,28.5v-2h-3a2.18,2.18,0,0,0-2-2v-3h-2v3h-3v-3h-2v3h-3v-3h-2v3h-3v-3h-2v3h-3v-3h-2v3a2.18,2.18,0,0,0-2,2h-3v2h3v3h-3v2h3v3h-3v2h3v3h-3v2h3v3h-3v2h3a2.18,2.18,0,0,0,2,2v3h2v-3h3v3h2v-3h3v3h2v-3h3v3h2v-3h3v3h2v-3a2.18,2.18,0,0,0,2-2h3v-2h-3v-3h3v-2h-3v-3h3v-2h-3v-3h3v-2h-3v-3Zm-5,19.82a.18.18,0,0,1-.18.18H26.68a.18.18,0,0,1-.18-.18V26.68a.18.18,0,0,1,.18-.18H48.32a.18.18,0,0,1,.18.18Z"/></g></g></svg>`,
-  AWS_RDS: `<svg xmlns="http://www.w3.org/2000/svg" xmlns:xlink="http://www.w3.org/1999/xlink" viewBox="0 0 75 75"><defs><style>.cls-1{fill:url(#BlueGradient);}.cls-2{fill:#fff;}</style><linearGradient id="BlueGradient" x1="-897.47" y1="-41.53" x2="-1003.54" y2="64.54" gradientTransform="translate(-913 49) rotate(180)" gradientUnits="userSpaceOnUse"><stop offset="0" stop-color="#2e27ad"/><stop offset="1" stop-color="#527fff"/></linearGradient></defs><title>Amazon-RDS</title><g id="Reference"><rect id="Blue_Gradient" data-name="Blue Gradient" class="cls-1" width="75" height="75"/><g id="Icon_Test" data-name="Icon Test"><path class="cls-2" d="M48.45,28.23c0-2.9-5.69-4.24-11-4.24s-11,1.34-11,4.24a1.49,1.49,0,0,0,.05.43V46.71c-.05,3,5.65,4.28,11,4.28s11-1.32,11-4.22V28.23ZM37.47,26c5.83,0,9,1.59,9,2.24s-3.15,2.24-9,2.24-9-1.58-9-2.24S31.64,26,37.47,26Zm9,20.78c0,.64-3.16,2.22-9,2.22s-9-1.58-9-2.22V43.62c2.11,1.15,5.65,1.7,9,1.7s6.86-.54,9-1.67Zm0-5.69c0,.64-3.15,2.24-9,2.24s-9-1.6-9-2.24H28.5V37.27c2.11,1.14,5.65,1.69,9,1.69s6.86-.54,9-1.67Zm0-6.37c0,.65-3.15,2.25-9,2.25s-9-1.6-9-2.25H28.5V30.82c2.13,1.13,5.63,1.65,9,1.65s6.91-.54,9-1.69Z"/><path class="cls-2" d="M15.91,60.51H22.5v2h-9a1,1,0,0,1-1-1v-9h2v6.57L22,51.59,23.4,53Z"/><path class="cls-2" d="M62.5,52.51v9a1,1,0,0,1-1,1h-9v-2h6.59L51.7,53.1l1.41-1.4,7.39,7.38V52.51Z"/><path class="cls-2" d="M62.5,13.51v9h-2v-6.6L53.11,23.3,51.7,21.89l7.38-7.38H52.5v-2h9A1,1,0,0,1,62.5,13.51Z"/><path class="cls-2" d="M23.4,22,22,23.4,14.5,15.91v6.58h-2v-9a1,1,0,0,1,1-1h9v2H15.91Z"/><path class="cls-2" d="M22.16,46.46c-6.11-2.2-9.61-5.56-9.61-9.21s3.5-7,9.61-9.21l.68,1.88c-5.19,1.87-8.29,4.61-8.29,7.33s3.1,5.46,8.29,7.33Z"/><path class="cls-2" d="M52.28,46.68l-.64-1.9c5.55-1.87,8.86-4.69,8.86-7.53s-3.31-5.66-8.86-7.54l.64-1.89C58.77,30,62.5,33.46,62.5,37.25S58.77,44.48,52.28,46.68Z"/></g></g></svg>`,
-  AZURE_VM: `<?xml version="1.0" encoding="utf-8"?>
-<!-- Generator: Adobe Illustrator 23.0.3, SVG Export Plug-In . SVG Version: 6.00 Build 0)  -->
-<svg version="1.1" id="Layer_1" xmlns="http://www.w3.org/2000/svg" xmlns:xlink="http://www.w3.org/1999/xlink" x="0px" y="0px"
-	 width="50px" height="50px" viewBox="0 0 50 50" enable-background="new 0 0 50 50" xml:space="preserve">
-<path fill="#7A7A7A" d="M33,38h-1.207H19.892h-0.621C20.92,43.999,18.705,44.859,9,44.859V48h12.342h9.011H42v-3.141
-	c-9.705,0-10.656-0.857-9.005-6.859"/>
-<path fill="#A0A1A2" d="M46.98,2H2.719C1.215,2,0.002,3.345,0.002,4.847v31.329c0,1.493,1.213,2.823,2.717,2.823H46.98
-	c1.501,0,3.021-1.33,3.021-2.823V4.847C50.001,3.341,48.481,2,46.98,2"/>
-<path opacity="0.2" fill="#FFFFFF" enable-background="new    " d="M47.011,2.003c-0.011,0-0.021-0.002-0.031-0.002H2.718
-	c-1.504,0-2.717,1.344-2.717,2.847v31.328C0.001,37.67,1.214,39,2.718,39h1.053L47.011,2.003z"/>
-<polygon fill="#59B4D9" points="46,6 46,35 4,35 4,6 "/>
-<rect x="9" y="45" fill="#A0A1A2" width="33" height="3"/>
-<path fill="#B8D432" d="M25.518,4.095c0,0.392-0.318,0.71-0.71,0.71c-0.393,0-0.709-0.318-0.709-0.71c0-0.393,0.316-0.71,0.709-0.71
-	C25.2,3.385,25.518,3.702,25.518,4.095"/>
-<path fill="#FFFFFF" d="M25.028,19.394c-0.045,0-0.091-0.014-0.134-0.038l-8.804-5.082c-0.081-0.048-0.132-0.137-0.132-0.231
-	c0-0.095,0.051-0.183,0.132-0.23l8.751-5.049c0.082-0.046,0.182-0.046,0.263,0l8.807,5.084c0.082,0.047,0.131,0.135,0.131,0.23
-	c0,0.096-0.049,0.183-0.131,0.23l-8.748,5.048C25.12,19.38,25.076,19.394,25.028,19.394"/>
-<path opacity="0.7" fill="#FFFFFF" enable-background="new    " d="M23.763,31.746c-0.05,0-0.095-0.012-0.134-0.036l-8.778-5.066
-	c-0.085-0.047-0.136-0.133-0.136-0.231V16.247c0-0.096,0.051-0.183,0.136-0.231c0.081-0.049,0.181-0.049,0.268,0l8.777,5.064
-	c0.078,0.05,0.13,0.137,0.13,0.233v10.166c0,0.097-0.052,0.183-0.13,0.231C23.852,31.734,23.806,31.746,23.763,31.746"/>
-<path opacity="0.4" fill="#FFFFFF" enable-background="new    " d="M26.248,31.746c-0.048,0-0.093-0.012-0.138-0.036
-	c-0.078-0.048-0.129-0.134-0.129-0.231V21.377c0-0.094,0.051-0.182,0.129-0.231l8.777-5.064c0.084-0.048,0.182-0.048,0.264,0
-	c0.084,0.047,0.135,0.135,0.135,0.23v10.101c0,0.098-0.051,0.184-0.135,0.231l-8.774,5.066
-	C26.34,31.734,26.294,31.746,26.248,31.746"/>
-</svg>
-`,
-  AZURE_VM_SCALE_SETS: `<?xml version="1.0" encoding="utf-8"?>
-<!-- Generator: Adobe Illustrator 17.0.1, SVG Export Plug-In . SVG Version: 6.00 Build 0)  -->
-<!DOCTYPE svg PUBLIC "-//W3C//DTD SVG 1.1//EN" "http://www.w3.org/Graphics/SVG/1.1/DTD/svg11.dtd">
-<svg version="1.1" id="Layer_1" xmlns="http://www.w3.org/2000/svg" xmlns:xlink="http://www.w3.org/1999/xlink" x="0px" y="0px"
-	 width="50px" height="50px" viewBox="0 0 50 50" enable-background="new 0 0 50 50" xml:space="preserve">
-<path fill="#A0A1A2" d="M35.723,0H2.066C0.923,0,0.001,1.021,0.001,2.162v23.439c0,1.134,0.922,2.144,2.066,2.144h33.656
-	c1.141,0,2.297-1.01,2.297-2.144V2.162C38.02,1.018,36.864,0,35.723,0"/>
-<path opacity="0.2" fill="#FFFFFF" d="M35.746,0.002c-0.008,0-0.016-0.002-0.024-0.002H2.066C0.922,0.001,0,1.021,0,2.163v23.438
-	c0,1.134,0.922,2.144,2.066,2.144h0.801L35.746,0.002z"/>
-<polygon fill="#59B4D9" points="35.129,24.823 2.882,24.823 2.882,2.922 35.129,2.875 "/>
-<path fill="#B8D432" d="M19.404,1.591c0,0.298-0.242,0.539-0.54,0.539c-0.299,0-0.539-0.241-0.539-0.539
-	c0-0.298,0.24-0.539,0.539-0.539C19.162,1.052,19.404,1.293,19.404,1.591"/>
-<path fill="#FFFFFF" d="M19.425,13.209c-0.035,0-0.069-0.011-0.102-0.029l-6.695-3.859c-0.061-0.036-0.1-0.104-0.1-0.175
-	c0-0.072,0.039-0.139,0.1-0.175l6.654-3.834c0.062-0.035,0.138-0.035,0.2,0l6.697,3.861c0.062,0.036,0.1,0.103,0.1,0.175
-	c0,0.073-0.037,0.139-0.1,0.175l-6.652,3.833C19.495,13.198,19.462,13.209,19.425,13.209"/>
-<path fill="#A0A1A2" d="M41.671,7.456H8.015c-1.144,0-2.066,1.021-2.066,2.162v23.439c0,1.134,0.922,2.144,2.066,2.144h33.656
-	c1.141,0,2.297-1.01,2.297-2.144V9.618C43.969,8.474,42.813,7.456,41.671,7.456"/>
-<path opacity="0.2" fill="#FFFFFF" d="M41.695,7.458c-0.008,0-0.016-0.002-0.024-0.002H8.015c-1.144,0-2.066,1.021-2.066,2.162
-	v23.438c0,1.134,0.922,2.144,2.066,2.144h0.801L41.695,7.458z"/>
-<polygon fill="#59B4D9" points="41.078,32.279 8.83,32.279 8.83,10.378 41.078,10.331 "/>
-<path fill="#B8D432" d="M25.352,9.047c0,0.298-0.242,0.539-0.54,0.539c-0.299,0-0.539-0.241-0.539-0.539
-	c0-0.298,0.24-0.539,0.539-0.539C25.11,8.508,25.352,8.748,25.352,9.047"/>
-<path fill="#FFFFFF" d="M25.374,20.664c-0.035,0-0.069-0.011-0.102-0.029l-6.695-3.859c-0.061-0.036-0.1-0.104-0.1-0.175
-	c0-0.072,0.039-0.139,0.1-0.175l6.654-3.834c0.062-0.035,0.138-0.035,0.2,0l6.697,3.861c0.062,0.036,0.1,0.103,0.1,0.175
-	c0,0.073-0.037,0.139-0.1,0.175l-6.652,3.833C25.443,20.654,25.411,20.664,25.374,20.664"/>
-<path fill="#7A7A7A" d="M36.636,42.698h-0.906h-8.965h-0.468c1.242,4.38-0.427,5.008-7.737,5.008V50h9.297h6.788h8.773v-2.293
-	C36.109,47.706,35.392,47.081,36.636,42.698"/>
-<path fill="#A0A1A2" d="M47.703,14.955H14.047c-1.144,0-2.066,1.021-2.066,2.162v23.439c0,1.134,0.922,2.144,2.066,2.144h33.656
-	c1.141,0,2.297-1.01,2.297-2.144V17.117C50,15.974,48.844,14.955,47.703,14.955"/>
-<path opacity="0.2" fill="#FFFFFF" d="M47.727,14.958c-0.008,0-0.016-0.002-0.024-0.002H14.046c-1.144,0-2.066,1.021-2.066,2.162
-	v23.438c0,1.134,0.922,2.144,2.066,2.144h0.801L47.727,14.958z"/>
-<polygon fill="#59B4D9" points="47.109,39.779 14.862,39.779 14.862,17.878 47.109,17.83 "/>
-<rect x="18.561" y="47.706" fill="#A0A1A2" width="24.858" height="2.294"/>
-<path fill="#B8D432" d="M31.384,16.546c0,0.298-0.242,0.539-0.54,0.539c-0.299,0-0.539-0.241-0.539-0.539
-	c0-0.298,0.24-0.539,0.539-0.539C31.142,16.007,31.384,16.248,31.384,16.546"/>
-<path fill="#FFFFFF" d="M31.405,28.164c-0.035,0-0.069-0.011-0.102-0.029l-6.695-3.859c-0.061-0.036-0.1-0.104-0.1-0.175
-	c0-0.072,0.039-0.139,0.1-0.175l6.654-3.834c0.062-0.035,0.138-0.035,0.2,0l6.697,3.861c0.062,0.036,0.1,0.103,0.1,0.175
-	c0,0.073-0.037,0.139-0.1,0.175l-6.652,3.833C31.475,28.153,31.442,28.164,31.405,28.164"/>
-<path opacity="0.7" fill="#FFFFFF" d="M30.443,37.544c-0.038,0-0.072-0.009-0.102-0.027l-6.675-3.847
-	c-0.065-0.036-0.104-0.101-0.104-0.175v-7.72c0-0.073,0.039-0.139,0.104-0.175c0.062-0.037,0.137-0.037,0.204,0l6.674,3.846
-	c0.059,0.038,0.099,0.104,0.099,0.177v7.72c0,0.074-0.039,0.139-0.099,0.175C30.511,37.535,30.476,37.544,30.443,37.544"/>
-<path opacity="0.4" fill="#FFFFFF" d="M32.333,37.544c-0.037,0-0.071-0.009-0.105-0.027c-0.059-0.036-0.098-0.102-0.098-0.175V29.67
-	c0-0.071,0.039-0.138,0.098-0.175l6.674-3.845c0.064-0.036,0.138-0.036,0.201,0c0.064,0.036,0.102,0.102,0.102,0.175v7.671
-	c0,0.074-0.039,0.14-0.102,0.175l-6.672,3.847C32.403,37.535,32.367,37.544,32.333,37.544"/>
-</svg>
-`,
-  AZURE_SQL_SERVER: `<svg xmlns="http://www.w3.org/2000/svg" class="fxs-portal-svg" id="FxSymbol0-0eb" role="presentation" viewBox="-0.5 0.5 50 50" focusable="false" xmlns:xlink="http://www.w3.org/1999/xlink" xmlns:svg="http://www.w3.org/2000/svg">
-<g>
-<title/>
-<path class="msportalfx-svg-c19" fill="#0072c6" d="M -0.5 7.288 v 36.111 c 0 3.749 8.392 6.789 18.743 6.789 v -42.9 H -0.5 Z"/>
-<path class="msportalfx-svg-c19" fill="#0072c6" d="M 17.986 50.187 h 0.257 c 10.351 0 18.743 -3.038 18.743 -6.788 V 7.288 h -19 v 42.899 Z"/>
-<path class="msportalfx-svg-c01" opacity="0.15" fill="#ffffff" d="M 17.986 50.187 h 0.257 c 10.351 0 18.743 -3.038 18.743 -6.788 V 7.288 h -19 v 42.899 Z"/>
-<path class="msportalfx-svg-c01" fill="#ffffff" d="M 36.986 7.288 c 0 3.749 -8.392 6.788 -18.743 6.788 S -0.5 11.037 -0.5 7.288 S 7.892 0.5 18.243 0.5 s 18.743 3.039 18.743 6.788"/>
-<path class="msportalfx-svg-c14" fill="#7fba00" d="M 33.154 6.897 c 0 2.475 -6.676 4.479 -14.911 4.479 S 3.331 9.372 3.331 6.897 c 0 -2.474 6.677 -4.479 14.912 -4.479 s 14.911 2.005 14.911 4.479"/>
-<path class="msportalfx-svg-c13" fill="#b8d432" d="M 30.03 9.634 c 1.952 -0.757 3.125 -1.705 3.125 -2.735 c 0 -2.475 -6.676 -4.48 -14.912 -4.48 c -8.235 0 -14.911 2.005 -14.911 4.48 c 0 1.03 1.173 1.978 3.125 2.735 c 2.726 -1.058 6.986 -1.741 11.786 -1.741 c 4.801 0 9.059 0.683 11.787 1.741"/>
-<path class="msportalfx-svg-c01" fill="#ffffff" d="M 12.29 31.354 c 0 1.122 -0.407 1.991 -1.221 2.607 c -0.814 0.616 -1.938 0.924 -3.373 0.924 c -1.221 0 -2.241 -0.22 -3.061 -0.66 v -2.64 c 0.946 0.803 1.988 1.205 3.126 1.205 c 0.55 0 0.975 -0.11 1.275 -0.33 s 0.45 -0.511 0.45 -0.875 c 0 -0.357 -0.144 -0.668 -0.433 -0.932 s -0.876 -0.605 -1.761 -1.023 c -1.804 -0.846 -2.706 -2.002 -2.706 -3.464 c 0 -1.061 0.393 -1.912 1.18 -2.553 c 0.786 -0.64 1.831 -0.961 3.134 -0.961 c 1.155 0 2.111 0.152 2.871 0.454 v 2.466 c -0.797 -0.55 -1.705 -0.825 -2.722 -0.825 c -0.511 0 -0.915 0.108 -1.212 0.325 a 1.026 1.026 0 0 0 -0.445 0.87 c 0 0.374 0.119 0.681 0.359 0.92 c 0.239 0.239 0.73 0.535 1.472 0.887 c 1.106 0.523 1.893 1.053 2.364 1.592 c 0.468 0.54 0.703 1.211 0.703 2.013 Z m 12.727 -2.672 c 0 1.391 -0.317 2.599 -0.949 3.621 c -0.633 1.023 -1.523 1.74 -2.672 2.153 l 3.431 3.176 h -3.464 l -2.45 -2.747 c -1.05 -0.038 -1.998 -0.316 -2.842 -0.833 c -0.844 -0.516 -1.496 -1.225 -1.955 -2.124 s -0.689 -1.902 -0.689 -3.007 c 0 -1.226 0.249 -2.319 0.746 -3.279 c 0.498 -0.96 1.197 -1.698 2.099 -2.215 c 0.902 -0.516 1.935 -0.775 3.102 -0.775 c 1.088 0 2.063 0.25 2.924 0.751 c 0.86 0.5 1.528 1.212 2.004 2.136 c 0.477 0.924 0.715 1.972 0.715 3.143 Z m -2.804 0.149 c 0 -1.199 -0.261 -2.146 -0.784 -2.842 s -1.237 -1.044 -2.145 -1.044 c -0.924 0 -1.663 0.349 -2.219 1.047 c -0.555 0.699 -0.833 1.628 -0.833 2.788 c 0 1.155 0.272 2.077 0.816 2.767 c 0.545 0.69 1.267 1.035 2.169 1.035 c 0.919 0 1.647 -0.334 2.186 -1.002 c 0.54 -0.667 0.81 -1.584 0.81 -2.749 Z m 11.803 5.848 h -7.044 V 22.851 h 2.664 v 9.667 h 4.38 Z"/>
-<path class="msportalfx-svg-c03" fill="#a0a1a2" d="M 34.755 49.584 l 2.123 0.916 l 0.224 -0.265 l 1.7 -1.999 l 1.719 0.01 l 1.973 2.254 l 2.135 -0.887 l -0.023 -0.347 l -0.185 -2.643 l 1.225 -1.234 l 2.957 0.168 l 0.897 -2.167 l -0.258 -0.229 l -1.96 -1.738 l 0.012 -1.756 l 2.205 -2.016 l -0.866 -2.179 l -0.342 0.025 l -2.588 0.186 l -1.206 -1.249 l 0.164 -3.02 l -2.124 -0.914 l -0.224 0.265 l -1.702 2 l -1.72 -0.011 l -1.975 -2.251 l -2.134 0.884 l 0.026 0.35 l 0.181 2.64 l -1.224 1.234 l -2.955 -0.168 l -0.896 2.167 l 0.259 0.229 l 1.959 1.736 l -0.012 1.756 l -2.207 2.016 l 0.867 2.179 l 0.343 -0.025 l 2.587 -0.186 l 1.208 1.251 l -0.163 3.018 Z m 0.608 -10.956 c 1.012 -2.452 3.778 -3.6 6.177 -2.565 c 2.402 1.036 3.525 3.858 2.513 6.308 c -1.013 2.451 -3.78 3.598 -6.179 2.565 c -2.4 -1.034 -3.524 -3.857 -2.511 -6.308"/>
-</g>
-</svg>`,
-  AZURE_SQL_DATABASE: `<svg xmlns="http://www.w3.org/2000/svg" class="fxs-portal-svg" id="FxSymbol0-0ea" role="presentation" viewBox="-0.5 0.5 50 50" focusable="false" xmlns:xlink="http://www.w3.org/1999/xlink" xmlns:svg="http://www.w3.org/2000/svg">
-<g>
-<title/>
-<path class="msportalfx-svg-c19" fill="#0072c6" d="M 5.757 7.288 v 36.111 c 0 3.749 8.392 6.789 18.743 6.789 v -42.9 H 5.757 Z"/>
-<path class="msportalfx-svg-c19" fill="#0072c6" d="M 24.243 50.187 h 0.257 c 10.351 0 18.743 -3.038 18.743 -6.788 V 7.288 h -19 v 42.899 Z"/>
-<path class="msportalfx-svg-c01" opacity="0.15" fill="#ffffff" d="M 24.243 50.187 h 0.257 c 10.351 0 18.743 -3.038 18.743 -6.788 V 7.288 h -19 v 42.899 Z"/>
-<path class="msportalfx-svg-c01" fill="#ffffff" d="M 43.243 7.288 c 0 3.749 -8.392 6.788 -18.743 6.788 S 5.757 11.037 5.757 7.288 S 14.149 0.5 24.5 0.5 s 18.743 3.039 18.743 6.788"/>
-<path class="msportalfx-svg-c14" fill="#7fba00" d="M 39.411 6.897 c 0 2.475 -6.676 4.479 -14.911 4.479 S 9.588 9.372 9.588 6.897 c 0 -2.474 6.677 -4.479 14.912 -4.479 s 14.911 2.005 14.911 4.479"/>
-<path class="msportalfx-svg-c13" fill="#b8d432" d="M 36.287 9.634 c 1.952 -0.757 3.125 -1.705 3.125 -2.735 c 0 -2.475 -6.676 -4.48 -14.912 -4.48 c -8.235 0 -14.911 2.005 -14.911 4.48 c 0 1.03 1.173 1.978 3.125 2.735 C 15.44 8.576 19.7 7.893 24.5 7.893 c 4.801 0 9.059 0.683 11.787 1.741"/>
-<path class="msportalfx-svg-c01" fill="#ffffff" d="M 18.547 32.354 c 0 1.122 -0.407 1.991 -1.221 2.607 c -0.814 0.616 -1.938 0.924 -3.373 0.924 c -1.221 0 -2.241 -0.22 -3.061 -0.66 v -2.64 c 0.946 0.803 1.988 1.205 3.126 1.205 c 0.55 0 0.975 -0.11 1.275 -0.33 s 0.45 -0.511 0.45 -0.875 c 0 -0.357 -0.144 -0.668 -0.433 -0.932 s -0.876 -0.605 -1.761 -1.023 c -1.804 -0.846 -2.706 -2.002 -2.706 -3.464 c 0 -1.061 0.393 -1.912 1.18 -2.553 c 0.786 -0.64 1.831 -0.961 3.134 -0.961 c 1.155 0 2.111 0.152 2.871 0.454 v 2.466 c -0.797 -0.55 -1.705 -0.825 -2.722 -0.825 c -0.511 0 -0.915 0.108 -1.212 0.325 a 1.026 1.026 0 0 0 -0.445 0.87 c 0 0.374 0.119 0.681 0.359 0.92 c 0.239 0.239 0.73 0.535 1.472 0.887 c 1.106 0.523 1.893 1.053 2.364 1.592 c 0.468 0.54 0.703 1.211 0.703 2.013 Z m 12.727 -2.672 c 0 1.391 -0.317 2.599 -0.949 3.621 c -0.633 1.023 -1.523 1.74 -2.672 2.153 l 3.431 3.176 H 27.62 l -2.45 -2.747 c -1.05 -0.038 -1.998 -0.316 -2.842 -0.833 c -0.844 -0.516 -1.496 -1.225 -1.955 -2.124 s -0.689 -1.902 -0.689 -3.007 c 0 -1.226 0.249 -2.319 0.746 -3.279 c 0.498 -0.96 1.197 -1.698 2.099 -2.215 c 0.902 -0.516 1.935 -0.775 3.102 -0.775 c 1.088 0 2.063 0.25 2.924 0.751 c 0.86 0.5 1.528 1.212 2.004 2.136 c 0.477 0.924 0.715 1.972 0.715 3.143 Z m -2.804 0.149 c 0 -1.199 -0.261 -2.146 -0.784 -2.842 s -1.237 -1.044 -2.145 -1.044 c -0.924 0 -1.663 0.349 -2.219 1.047 c -0.555 0.699 -0.833 1.628 -0.833 2.788 c 0 1.155 0.272 2.077 0.816 2.767 c 0.545 0.69 1.267 1.035 2.169 1.035 c 0.919 0 1.647 -0.334 2.186 -1.002 c 0.54 -0.667 0.81 -1.584 0.81 -2.749 Z m 11.803 5.848 h -7.044 V 23.851 h 2.664 v 9.667 h 4.38 Z"/>
-</g>
-</svg>`,
-  SEARCH: `<?xml version="1.0"?><svg fill="#bbbbc9" xmlns="http://www.w3.org/2000/svg" viewBox="0 0 24 24" width="48px" height="48px">    <path d="M 9 2 C 5.1458514 2 2 5.1458514 2 9 C 2 12.854149 5.1458514 16 9 16 C 10.747998 16 12.345009 15.348024 13.574219 14.28125 L 14 14.707031 L 14 16 L 20 22 L 22 20 L 16 14 L 14.707031 14 L 14.28125 13.574219 C 15.348024 12.345009 16 10.747998 16 9 C 16 5.1458514 12.854149 2 9 2 z M 9 4 C 11.773268 4 14 6.2267316 14 9 C 14 11.773268 11.773268 14 9 14 C 6.2267316 14 4 11.773268 4 9 C 4 6.2267316 6.2267316 4 9 4 z"/></svg>`,
-  ARROW_TOP: `<svg xmlns="http://www.w3.org/2000/svg" width="24" height="24" viewBox="0 0 24 24"><path d="M0 16.67l2.829 2.83 9.175-9.339 9.167 9.339 2.829-2.83-11.996-12.17z"/></svg>`,
-  ARROW_BOTTOM: `<svg xmlns="http://www.w3.org/2000/svg" width="24" height="24" viewBox="0 0 24 24"><path d="M0 7.33l2.829-2.83 9.175 9.339 9.167-9.339 2.829 2.83-11.996 12.17z"/></svg>`,
-  INPUT_ERROR: `<svg xmlns="http://www.w3.org/2000/svg" xmlns:xlink="http://www.w3.org/1999/xlink" viewBox="0 0 96 96" style="enable-background:new 0 0 96 96;" xml:space="preserve"><style type="text/css">.st0{fill:#E74C3C;}</style><g id="XMLID_7_"><path id="XMLID_13_" class="st0" d="M45.1,56.8h5.8v5.8h-5.8V56.8z M45.1,33.4h5.8v17.5h-5.8V33.4z M48,18.8 c-16.1,0-29.2,13.1-29.2,29.2s13,29.2,29.2,29.2S77.2,64.1,77.2,48S64.1,18.8,48,18.8z M48,71.3c-12.9,0-23.3-10.4-23.3-23.3 S35.1,24.7,48,24.7S71.3,35.1,71.3,48S60.9,71.3,48,71.3z"/></g></svg>`,
-  INPUT_SUCCESS: `<svg xmlns="http://www.w3.org/2000/svg"><path d="M15.812 0.896899C15.5652 0.655586 15.1695 0.659961 14.9282 0.906774L4.31249 11.7623L1.07187 8.44845C0.830555 8.20163 0.434774 8.19726 0.188024 8.43857C-0.0587576 8.67988 -0.0631951 9.0756 0.178149 9.32238L3.86559 13.0932C3.98118 13.2115 4.14124 13.2813 4.31228 13.2813C4.47549 13.2813 4.63775 13.2176 4.75931 13.0932L15.8218 1.78071C16.0632 1.53393 16.0588 1.13824 15.812 0.896899Z" fill="#27AE60"/></svg>`,
-  CLOSE: `<svg width="16" height="16" viewBox="0 0 16 16" fill="none" xmlns="http://www.w3.org/2000/svg">
-  <path d="M10.64 8.0005L15.5898 3.05069C16.1367 2.50379 16.1367 1.61712 15.5898 1.07107L14.9299 0.411202C14.3828 -0.135871 13.4962 -0.135871 12.9501 0.411202L8.0005 5.36084L3.05069 0.410177C2.50379 -0.136726 1.61712 -0.136726 1.07107 0.410177L0.410177 1.07005C-0.136726 1.61712 -0.136726 2.50379 0.410177 3.04984L5.36084 8.0005L0.411202 12.9501C-0.135871 13.4972 -0.135871 14.3839 0.411202 14.9299L1.07107 15.5898C1.61798 16.1367 2.50464 16.1367 3.05069 15.5898L8.0005 10.64L12.9501 15.5898C13.4972 16.1367 14.3839 16.1367 14.9299 15.5898L15.5898 14.9299C16.1367 14.3828 16.1367 13.4962 15.5898 12.9501L10.64 8.0005Z" fill={fill || 'black'}/>
-  </svg>
-  `,
-  CLOSE_WHITE: `<svg width="16" height="16" viewBox="0 0 16 16" fill="none" xmlns="http://www.w3.org/2000/svg">
-  <path d="M9.41502 7.99983L15.7068 1.70803C16.0977 1.31746 16.0977 0.683746 15.7068 0.293177C15.3159 -0.0977257 14.6829 -0.0977257 14.292 0.293177L8.00017 6.58498L1.70803 0.293177C1.31713 -0.0977257 0.68408 -0.0977257 0.293177 0.293177C-0.0977257 0.683746 -0.0977257 1.31746 0.293177 1.70803L6.58531 7.99983L0.293177 14.2916C-0.0977257 14.6822 -0.0977257 15.3159 0.293177 15.7065C0.488629 15.9016 0.744783 15.9993 1.0006 15.9993C1.25643 15.9993 1.51258 15.9016 1.70803 15.7062L8.00017 9.41435L14.292 15.7062C14.4874 15.9016 14.7436 15.9993 14.9994 15.9993C15.2552 15.9993 15.5114 15.9016 15.7068 15.7062C16.0977 15.3156 16.0977 14.6819 15.7068 14.2913L9.41502 7.99983Z" fill="#FCFCFD"/>
-  </svg>`,
-  FILTER: `<svg width="16" height="16" viewBox="0 0 16 16" fill="none" xmlns="http://www.w3.org/2000/svg">
-  <path d="M0 0.92218V3.15637L6.2745 9.43087V14.2151L9.7255 15.0779V9.43084L16 3.15634V0.92218H0Z" fill="#161632"/>
-  </svg>
-  `,
-  CLOSE_SIDEBAR: `<svg version="1.1" xmlns="http://www.w3.org/2000/svg" xmlns:xlink="http://www.w3.org/1999/xlink" x="0px" y="0px" width="24px"
-	 height="24px" viewBox="0 0 24 24" style="enable-background:new 0 0 24 24;" xml:space="preserve">
-<g id="Icons" style="opacity:0.75;">
-	<g id="close">
-		<polygon id="x" style="fill-rule:evenodd;clip-rule:evenodd;" points="18.717,6.697 17.303,5.283 12,10.586 6.697,5.283
-			5.283,6.697 10.586,12 5.283,17.303 6.697,18.717 12,13.414 17.303,18.717 18.717,17.303 13.414,12 		"/>
-	</g>
-</g>
-<g id="Guides" style="display:none;">
-</g>
-</svg>`,
-  OPEN_SIDEBAR: `<svg height="32px" id="Layer_1" style="enable-background:new 0 0 32 32;" version="1.1" viewBox="0 0 32 32" width="32px" xml:space="preserve" xmlns="http://www.w3.org/2000/svg" xmlns:xlink="http://www.w3.org/1999/xlink"><path d="M4,10h24c1.104,0,2-0.896,2-2s-0.896-2-2-2H4C2.896,6,2,6.896,2,8S2.896,10,4,10z M28,14H4c-1.104,0-2,0.896-2,2  s0.896,2,2,2h24c1.104,0,2-0.896,2-2S29.104,14,28,14z M28,22H4c-1.104,0-2,0.896-2,2s0.896,2,2,2h24c1.104,0,2-0.896,2-2  S29.104,22,28,22z"/></svg>`,
-  CROSS_LIGHT: `<svg width="12" height="12" viewBox="0 0 12 12" fill="none" xmlns="http://www.w3.org/2000/svg">
-<path d="M7.98 6.00038L11.6924 2.28802C12.1025 1.87784 12.1025 1.21284 11.6924 0.803307L11.1975 0.308402C10.7872 -0.101903 10.1222 -0.101903 9.71262 0.308402L6.00038 4.02064L2.28802 0.307633C1.87784 -0.102544 1.21284 -0.102544 0.803307 0.307633L0.307633 0.802538C-0.102544 1.21284 -0.102544 1.87784 0.307633 2.28738L4.02064 6.00038L0.308402 9.71262C-0.101903 10.1229 -0.101903 10.7879 0.308402 11.1975L0.803307 11.6924C1.21348 12.1025 1.87848 12.1025 2.28802 11.6924L6.00038 7.98L9.71262 11.6924C10.1229 12.1025 10.7879 12.1025 11.1975 11.6924L11.6924 11.1975C12.1025 10.7872 12.1025 10.1222 11.6924 9.71262L7.98 6.00038Z" fill="#FCFCFD"/>
-</svg>`,
-  CROSS_DARK: `<svg width="12" height="12" viewBox="0 0 12 12" fill="none" xmlns="http://www.w3.org/2000/svg">
-<path d="M7.98 6.00038L11.6924 2.28802C12.1025 1.87784 12.1025 1.21284 11.6924 0.803307L11.1975 0.308402C10.7872 -0.101903 10.1222 -0.101903 9.71262 0.308402L6.00038 4.02064L2.28802 0.307633C1.87784 -0.102544 1.21284 -0.102544 0.803307 0.307633L0.307633 0.802538C-0.102544 1.21284 -0.102544 1.87784 0.307633 2.28738L4.02064 6.00038L0.308402 9.71262C-0.101903 10.1229 -0.101903 10.7879 0.308402 11.1975L0.803307 11.6924C1.21348 12.1025 1.87848 12.1025 2.28802 11.6924L6.00038 7.98L9.71262 11.6924C10.1229 12.1025 10.7879 12.1025 11.1975 11.6924L11.6924 11.1975C12.1025 10.7872 12.1025 10.1222 11.6924 9.71262L7.98 6.00038Z" fill="#161632"/>
-</svg>`,
-  HAMBURGER_DARK: `<svg width="48" height="48" viewBox="0 0 24 24" fill="none" xmlns="http://www.w3.org/2000/svg">
-<path d="M7 6H20V8H7V6Z" fill="#161632"/>
-<path d="M7 11H20V13H7V11Z" fill="#161632"/>
-<path d="M7 16H20V18H7V16Z" fill="#161632"/>
-<path d="M4 6H6V8H4V6Z" fill="#161632"/>
-<path d="M4 11H6V13H4V11Z" fill="#161632"/>
-<path d="M4 16H6V18H4V16Z" fill="#161632"/>
-</svg>
-`,
-  SUCCESS: `<svg width="24" height="24" viewBox="0 0 24 24" fill="none" xmlns="http://www.w3.org/2000/svg">
-  <path d="M15.9052 9.59476C15.6127 9.30189 15.1372 9.30189 14.8447 9.59476L11.625 12.8141L10.2802 11.4698C9.98772 11.1769 9.51222 11.1769 9.21972 11.4698C8.92684 11.7626 8.92684 12.2374 9.21972 12.5303L11.0947 14.4053C11.241 14.5519 11.433 14.625 11.625 14.625C11.817 14.625 12.009 14.5519 12.1552 14.4053L15.9052 10.6553C16.1981 10.3624 16.1981 9.88764 15.9052 9.59476Z" fill="white"/>
-  <path d="M19.3333 11.3333C18.9653 11.3333 18.6667 11.632 18.6667 12C18.6667 15.676 15.676 18.6667 12 18.6667C8.324 18.6667 5.33333 15.676 5.33333 12C5.33333 8.324 8.324 5.33333 12 5.33333C13.7897 5.33333 15.4697 6.03267 16.731 7.30267C16.99 7.56433 17.4123 7.56567 17.6737 7.306C17.935 7.04667 17.9363 6.62467 17.677 6.36333C16.1637 4.83933 14.1473 4 12 4C7.58867 4 4 7.58867 4 12C4 16.4113 7.58867 20 12 20C16.4113 20 20 16.4113 20 12C20 11.632 19.7013 11.3333 19.3333 11.3333Z" fill="white"/>
-  </svg>`,
-  DANGER: `<svg width="24" height="24" viewBox="0 0 24 24" fill="none" xmlns="http://www.w3.org/2000/svg">
-  <path d="M12.0004 16.9626C11.6552 16.9626 11.3751 16.6829 11.3751 16.3376C11.3751 15.9924 11.6548 15.7126 12 15.7126H12.0004C12.3455 15.7126 12.6254 15.9924 12.6254 16.3376C12.6254 16.6829 12.3455 16.9626 12.0004 16.9626ZM12.625 14.1501V10.2751C12.625 9.92993 12.3452 9.65015 12 9.65015C11.6548 9.65015 11.375 9.92993 11.375 10.2751V14.1501C11.375 14.4954 11.6548 14.7751 12 14.7751C12.3452 14.7751 12.625 14.4954 12.625 14.1501ZM20 18.5392V18.447C20 18.2975 19.962 18.1492 19.8901 18.0181L18.8038 16.0371C18.6378 15.7345 18.2579 15.6237 17.9552 15.7897C17.6526 15.9556 17.5417 16.3356 17.7078 16.6382L18.5541 18.1814H5.44592L12 6.2301L16.2854 14.0444C16.4513 14.347 16.8313 14.4579 17.1339 14.2919C17.4365 14.1259 17.5474 13.746 17.3813 13.4434L12.7822 5.05688C12.6256 4.77124 12.3258 4.59375 12 4.59375C11.6742 4.59375 11.3744 4.77124 11.2178 5.05688L4.10986 18.0181C4.03796 18.1492 4 18.2975 4 18.447V18.5392C4 19.0311 4.40027 19.4314 4.89221 19.4314H19.1078C19.5997 19.4314 20 19.0311 20 18.5392Z" fill="white"/>
-  </svg>`,
-  NOTIFY: `<svg width="24" height="24" viewBox="0 0 24 24" fill="none" xmlns="http://www.w3.org/2000/svg">
-  <path d="M12 8.59375C12.3452 8.59375 12.625 8.31393 12.625 7.96875C12.625 7.62357 12.3452 7.34375 12 7.34375C11.6548 7.34375 11.375 7.62357 11.375 7.96875C11.375 8.31393 11.6548 8.59375 12 8.59375Z" fill="white"/>
-  <path d="M17.6569 6.34313C16.1458 4.83216 14.1369 4 12 4C9.86312 4 7.85416 4.83216 6.34313 6.34313C4.83216 7.85416 4 9.86312 4 12C4 14.1369 4.83216 16.1458 6.34313 17.6569C7.85416 19.1678 9.86312 20 12 20C13.4629 20 14.8943 19.6013 16.1394 18.8471C16.4347 18.6682 16.529 18.2839 16.3502 17.9887C16.1713 17.6934 15.787 17.5991 15.4918 17.7779C14.442 18.4139 13.2346 18.75 12 18.75C8.27803 18.75 5.25 15.722 5.25 12C5.25 8.27803 8.27803 5.25 12 5.25C15.722 5.25 18.75 8.27803 18.75 12C18.75 13.3309 18.3542 14.6264 17.6054 15.7464C17.4136 16.0333 17.4907 16.4215 17.7776 16.6133C18.0646 16.8051 18.4527 16.7281 18.6446 16.4411C19.5313 15.1148 20 13.5791 20 12C20 9.86312 19.1678 7.85416 17.6569 6.34313Z" fill="white"/>
-  <path d="M12 9.84375C11.6548 9.84375 11.375 10.1236 11.375 10.4688V16.0312C11.375 16.3764 11.6548 16.6562 12 16.6562C12.3452 16.6562 12.625 16.3764 12.625 16.0312V10.4688C12.625 10.1236 12.3452 9.84375 12 9.84375Z" fill="white"/>
-  </svg>`,
-  PLUS: `<svg width="24" height="24" viewBox="0 0 24 24" fill="none" xmlns="http://www.w3.org/2000/svg">
-  <path d="M12 4C7.58853 4 4 7.58853 4 12C4 16.4115 7.58853 20 12 20C16.4115 20 20 16.4108 20 12C20 7.58916 16.4115 4 12 4ZM12 18.7607C8.27266 18.7607 5.23934 15.728 5.23934 12C5.23934 8.27203 8.27266 5.23934 12 5.23934C15.7273 5.23934 18.7607 8.27203 18.7607 12C18.7607 15.728 15.728 18.7607 12 18.7607Z" fill="#161632"/>
-  <path d="M15.0977 11.3244H12.619V8.84576C12.619 8.5037 12.342 8.22607 11.9993 8.22607C11.6566 8.22607 11.3796 8.5037 11.3796 8.84576V11.3244H8.90094C8.55825 11.3244 8.28125 11.6021 8.28125 11.9441C8.28125 12.2862 8.55825 12.5638 8.90094 12.5638H11.3796V15.0425C11.3796 15.3846 11.6566 15.6622 11.9993 15.6622C12.342 15.6622 12.619 15.3846 12.619 15.0425V12.5638H15.0977C15.4404 12.5638 15.7174 12.2862 15.7174 11.9441C15.7174 11.6021 15.4404 11.3244 15.0977 11.3244Z" fill="#161632"/>
-  </svg>`,
-  MENU: `<svg width="20" height="20" viewBox="0 0 20 20" fill="none" xmlns="http://www.w3.org/2000/svg"><path d="M5.83301 5H16.6663V6.66667H5.83301V5Z" fill="#161632"/><path d="M5.83301 9.16699H16.6663V10.8337H5.83301V9.16699Z" fill="#161632"/><path d="M5.83301 13.333H16.6663V14.9997H5.83301V13.333Z" fill="#161632"/><path d="M3.33301 5H4.99967V6.66667H3.33301V5Z" fill="#161632"/><path d="M3.33301 9.16699H4.99967V10.8337H3.33301V9.16699Z" fill="#161632"/><path d="M3.33301 13.333H4.99967V14.9997H3.33301V13.333Z" fill="#161632"/></svg>
-`,
-};
