--- src/utils/svgicons.tsx
@@ -1,27 +1,0 @@
-import React from "react";
-import { awsIcons } from "./icons/aws-resource";
-import { cloudIcons } from "./icons/cloud-provider";
-import { klarityIcons } from "./icons/klarity-resource";
-import { menuIcons } from "./icons/menu";
-import { osComponentsIcons } from "./icons/os-components";
-import { osProviderIcons } from "./icons/os-provider";
-import { uiIcons } from "./icons/ui";
-
-export const getViewBox = () => "0 0 24 24";
-
-export const paths = {
-  ...awsIcons,
-  ...cloudIcons,
-  ...klarityIcons,
-  ...menuIcons,
-  ...uiIcons,
-  ...osProviderIcons,
-  ...osComponentsIcons,
-  default: <path />,
-} as const;
-
-export const getPath = (name: PathName) => {
-  return paths[name] ?? paths["default"];
-};
-
-export type PathName = keyof typeof paths;
